+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Library:
 6. [fix] 修复导入 TemplateMatchPrefab 与 OcrPrefab 时会自动导入开发依赖 rich 的问题。
 7. [feat] SendMessageImpl 现在可选开启运行时阻止系统休眠的功能，默认开启。
 8. [fix] 修复 ScrcpyImpl 会调用 PATH 里的 adb，若系统未安装 adb 从而抛出异常的问题。
+9. [feat] MuMu12V5Host 新增 `check_app_keptlive` 可用于检查是否开启 APP 后台保活模式。
 
 Devtool：
 1. [feat] 新增最近打开文件功能。